+++ lua/me/plugins/dap/nvim_dap.lua
@@ -5,6 +5,11 @@ return {
 	},
 	opts = function()
 		local dap = require("dap")
+		local pickers = require("telescope.pickers")
+		local finders = require("telescope.finders")
+		local actions = require("telescope.actions")
+		local action_state = require("telescope.actions.state")
+		local conf = require("telescope.config").values
 
 		-- Setup C#
 		dap.adapters.coreclr = {
@@ -16,14 +21,44 @@ return {
 		dap.configurations.cs = {
 			{
 				type = "coreclr",
+				name = "[DAP dotnet] Launch",
+				request = "launch",
+				stopAtEntry = false,
+				env = {
+					ASPNETCORE_ENVIRONMENT = "Development",
+				},
+				cwd = "${workspaceFolder}",
+				args = {},
+				program = function()
+                    os.execute("dotnet build -c Debug")
+
+					return coroutine.create(function(co)
+						local opts = {}
+						pickers
+							.new(opts, {
+								prompt_title = "Launch",
+								finder = finders.new_oneshot_job({ "find", ".", "-type", "f", "-path", "*.dll" }, {}),
+								sorter = conf.generic_sorter(opts),
+								attach_mappings = function(buffer_number)
+									actions.select_default:replace(function()
+										actions.close(buffer_number)
+										coroutine.resume(
+											co,
+											"${workspaceFolder}/" .. action_state.get_selected_entry()[1]
+										)
+									end)
+									return true
+								end,
+							})
+							:find()
+					end)
+				end,
+			},
+			{
+				type = "coreclr",
 				name = "[DAP dotnet] Attach to process",
 				request = "attach",
 				processId = function()
-					local pickers = require("telescope.pickers")
-					local finders = require("telescope.finders")
-					local actions = require("telescope.actions")
-					local action_state = require("telescope.actions.state")
-					local conf = require("telescope.config").values
 					return coroutine.create(function(co)
 						local opts = {}
 						local user = vim.fn.trim(vim.fn.system("whoami"))